pyproject.toml
[project]
name = "setup-agent"
version = "0.1.0"
description = "LLM Powered open source project setup agent"
authors = [
    {name = "Setup-Agent Contributors"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "docker>=6.0.0",
    "litellm>=1.73.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
sag = "main:cli"

[tool.hatch.build.targets.wheel]
packages = ["agent", "config", "docker_orch", "tools"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
